Add shell binary
We use [linenoise](https://github.com/antirez/linenoise/) lib for line
editing support (without history or completions for now).

As _linenoise_ is blocking and we need non-blocking output for JS
callbacks we use a design based on 2 threads:

- the main thread handle the classic JS runtime work and listen user
commands comming from _linenoise_ on unix socket, using the same I/O
event loop as JS (so events can either be a JS callback or a user command)
- linenoise is launched in a dedicated thread and sends user input
commands to the main thread by writing on the unix socket

The build system has been changed to split main between the 2
executables (bench and shell) and the tests.

It was also refactor to allow better integration as a lib.

Signed-off-by: Francis Bouvier <francis.bouvier@gmail.com>