Rename postInit in postJSObjectInst
Rationale: this feature has nothing to do with the Type init.
The idea is to get/set values on the JS object itself after
its instantiation.
Thoses actions are by definition dynamic and do not affect
the underling Type.
By consequence the corresponding Zig object will not have
those values.

Signed-off-by: Francis Bouvier <francis@lightpanda.io>