Simplify reflect.lookupPrototype
Switch from stable sort which was copied from std, to std.mem.sortUnstable.
Remove block.zig. This also has the benefit of working as-is in Zig 0.14,
whereas block.zig / std.mem.sort doesn't (and I can't figure out why).

lookupPrototype is now done in a single pass (post-sort). Previous, the initial
pass was only really used to set proto_index which the sorter could use (and,
in doing so, invalidate proto_index). But, as far as I can tell, proto_index
wasn't needed to sort. All you need to know is whether proto_T is null or not
and, if both A and B have a proto_T, whether a.proto_T == b.T.

Also optimized lookupPrototype to only scan structs before the current struct,
as should be guaranteed by the sort.