Add RemoteObject subtype
Allow Zig structures to define a `sub_type` which will be included in the
RemoteObject's subtype field (and will generate an empty description field)

External PersitedObjects now wrap our own ExternalEntry wrapper, which includes
a pointer to the Zig instance (what we used to have before). This approach
allows us to associate metadata with a v8.Value. We currently add the subtype
if available.